Moved sliders to their own component
