+++ src/components/Technologies/TechSlider.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { BsChevronCompactLeft, BsChevronCompactRight } from "react-icons/bs";
+import { motion, AnimatePresence } from "framer-motion";
 import Tech from "../../assets/Technology Images/smallweb.png";
 import Amazon from "../../assets/Technology slider/amazon.png";
 import Airbnb from "../../assets/Technology slider/airbnb.png";
@@ -25,14 +26,15 @@ const TechSlider = () => {
     };
   }, []);
 
-  useEffect(() => {
-    const interval = setInterval(() => {
-      handleNextSlide();
-    }, 4000);
+  const handleAutomaticSlide = () => {
+    setCurrentSlide((prevSlide) =>
+      prevSlide >= images.length - 2 ? 0 : prevSlide + 2
+    );
+  };
 
-    return () => {
-      clearInterval(interval);
-    };
+  useEffect(() => {
+    const interval = setInterval(handleAutomaticSlide, 4000);
+    return () => clearInterval(interval);
   });
 
   const handlePrevSlide = () => {
@@ -52,48 +54,55 @@ const TechSlider = () => {
     const secondImage = images[currentSlide + 1];
 
     return (
-      <div
-        className="slideshow-image flex justify-center items-center"
-        style={{
-          width: "100%",
-          height: "30%",
-          position: "absolute",
-        }}
-      >
-        <div
-          className="slideshow-image-item mt-2"
+      <AnimatePresence mode="wait">
+        <motion.div
+          key={currentSlide}
+          className="slideshow-image flex justify-center items-center"
           style={{
-            flex: "1",
-            maxWidth: "40%",
-            height: "100%",
-            backgroundSize: "contain",
-            backgroundPosition: "center",
-            backgroundRepeat: "no-repeat",
-            backgroundImage: `url(${firstImage})`,
+            width: "100%",
+            height: "30%",
+            position: "absolute",
           }}
-        ></div>
-        <div
-          className="slideshow-image-item mt-2"
-          style={{
-            flex: "1",
-            maxWidth: "40%",
-            height: "100%",
-            backgroundSize: "contain",
-            backgroundPosition: "center",
-            backgroundRepeat: "no-repeat",
-            backgroundImage: `url(${secondImage})`,
-          }}
-        ></div>
-      </div>
+          initial={{ opacity: 0, x: 100 }}
+          animate={{ opacity: 1, x: 0 }}
+          exit={{ opacity: 0, x: -100 }}
+          transition={{ duration: 0.5 }}
+        >
+          <motion.div
+            className="slideshow-image-item mt-2"
+            style={{
+              flex: "1",
+              maxWidth: "40%",
+              height: "100%",
+              backgroundSize: "contain",
+              backgroundPosition: "center",
+              backgroundRepeat: "no-repeat",
+              backgroundImage: `url(${firstImage})`,
+            }}
+          ></motion.div>
+          <motion.div
+            className="slideshow-image-item mt-2"
+            style={{
+              flex: "1",
+              maxWidth: "40%",
+              height: "100%",
+              backgroundSize: "contain",
+              backgroundPosition: "center",
+              backgroundRepeat: "no-repeat",
+              backgroundImage: `url(${secondImage})`,
+            }}
+          ></motion.div>
+        </motion.div>
+      </AnimatePresence>
     );
   };
-
   return (
     <div className="max-w-[1400px] mx-auto p-5 flex flex-col items-center">
       {isSmallScreen ? (
         <>
           <h1 className="font-semibold text-2xl sm:text-xl lg:text-4xl text-center mt-8 sm:mt-10 mb-4 sm:mb-6 lg:mb-8">
-            <span className="text-amber-500">Web Technologies</span> We Work With
+            <span className="text-amber-500">Web Technologies</span> We Work
+            With
           </h1>
           <div className="max-w-[1400px] h-[250px] w-full m-auto py-16 relative group">
             {/* Left Arrow */}
@@ -124,7 +133,12 @@ const TechSlider = () => {
             With
           </h1>
           <div className="flex justify-center items-center">
-            <img src={Tech} alt="Tech" className="pt-10 pb-10" style={{width: "1034", height: "191"}}/>
+            <img
+              src={Tech}
+              alt="Tech"
+              className="pt-10 pb-10"
+              style={{ width: "1034", height: "191" }}
+            />
           </div>
         </div>
       )}
